requirements.txt
fastapi
uvicorn
pydantic
python-multipart
google-genai
python-dotenv
requests
pywebview
pillow
cryptography